LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update-bytecheck.https-expected.txt


Harness Error (TIMEOUT), message = null

FAIL Test(cors: false, main: default, imported: default) assert_array_equals: property 2, expected null but got object "[object ServiceWorker]"
PASS Test(cors: false, main: default, imported: time) 
PASS Test(cors: false, main: time, imported: default) 
PASS Test(cors: false, main: time, imported: time) 
TIMEOUT Test(cors: true, main: default, imported: default) Test timed out
NOTRUN Test(cors: true, main: default, imported: time) 
NOTRUN Test(cors: true, main: time, imported: default) 
NOTRUN Test(cors: true, main: time, imported: time) 
